fix user printf for uint64

+++ user/printf.c
@@ -11,7 +11,7 @@ static void putc(int fd, char c) { write(fd, &c, 1); }
 static void printint(int fd, long long xx, int base, int sgn) {
   char buf[20];
   int i, neg;
-  uint x;
+  unsigned long long x;
 
   neg = 0;
   if (sgn && xx < 0) {
